pyproject.toml
[project]
name = "dream-factory-evals"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Hamza Farhan", email = "thehamza96@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.5",
    "langfuse>=2.60.2",
    "logfire>=3.11.0",
    "loguru>=0.7.3",
    "mcp[cli]>=1.5.0",
    "pandas>=2.2.3",
    "polars>=1.26.0",
    "pyarrow>=19.0.1",
    "pydantic-ai>=0.0.50.dev3",
    "python-dotenv>=1.1.0",
]

[project.scripts]
dream-factory-evals = "dream_factory_evals:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.sdist]
exclude = ["**/", "!src/"]

[tool.uv.sources]
pydantic-ai = {git = "https://github.com/pydantic/pydantic-ai.git", branch = "main"}

[dependency-groups]
dev = [
    "pytest>=8.3.5",
]
